Refactor with using form & directory
