+++ src/features/todo/components/TodoHeader.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const TodoHeader = () => <header className={'font-bold-32 text-black my-5 self-start'}>Hi! 🤚 Catze!</header>
+
+export default React.memo(TodoHeader);
