+++ src/pages/index.tsx
@@ -1,1 +1,1 @@
-export { default } from '@/features/todo/pages/Landing/LandingPage';
+export { default } from '@/features/todo/pages/LandingPage';
